+++ assets/javascripts/kitten/components/comments/comment-form/stories.js
@@ -1,37 +1,73 @@
 import React from 'react'
-import { StyleRoot } from 'radium'
 import { storiesOf } from '@storybook/react'
 import { withInfo } from '@storybook/addon-info'
 import { withKnobs, text, object, boolean } from '@storybook/addon-knobs'
 import { CommentForm } from './index'
 import { Grid, GridCol } from '../../../components/grid/grid'
 
-storiesOf('Comment/CommentForm', module)
+const info = {
+  text: `
+    # CommentForm
+    &nbsp;
+
+    ## Import
+    ~~~js
+    import { CommentForm } from '@kisskissbankbank/kitten/src/components/comments/comment-form'
+    ~~~
+
+    ## Usage
+    &nbsp;
+
+    #### Default
+    ~~~js
+    <CommentForm />
+    ~~~
+
+    ### With avatar
+    ~~~js
+    <CommentForm
+      avatarImgProps={{
+        src: '…'
+        alt: '…'
+      }}
+    />
+    ~~~
+
+    ### With commentButton
+    ~~~js
+    <CommentForm commentButton="Button" />
+    ~~~
+
+    ### With error
+    ~~~js
+    <CommentForm error errorMessage="…" />
+    ~~~
+  `,
+  header: false,
+  propTables: false,
+}
+
+storiesOf('Comments/CommentForm', module)
   .addDecorator(withKnobs)
-  .add(
-    'default',
-    withInfo('common info')(() => (
-      <StyleRoot>
-        <Grid>
-          <GridCol offset="1" col="8">
-            <CommentForm
-              avatarImgProps={object('Src/Alt', {
-                src: 'https://placehold.it/80x80/caf4fe/caf4fe',
-                alt: '',
-              })}
-              placeholder={text(
-                'placeholder',
-                'Leave a comment on the project',
-              )}
-              commentButton={text('Comment button', 'Comment')}
-              error={boolean('Is error', false)}
-              errorMessage={text(
-                'Error message',
-                'Cum sociis natoque penatibus et magnis',
-              )}
-            />
-          </GridCol>
-        </Grid>
-      </StyleRoot>
-    )),
-  )
+  .addDecorator(withInfo)
+  .add('Comment form', () => {
+    return (
+      <Grid style={{ marginTop: '5em' }}>
+        <GridCol offset="1" col="8">
+          <CommentForm
+            avatarImgProps={object('Src/Alt', {
+              src: 'https://placehold.it/80x80/caf4fe/caf4fe',
+              alt: '',
+            })}
+            placeholder={text('placeholder', 'Leave a comment on the project')}
+            commentButton={text('Comment button', 'Comment')}
+            error={boolean('Is error', false)}
+            errorMessage={text(
+              'Error message',
+              'Cum sociis natoque penatibus et magnis',
+            )}
+          />
+        </GridCol>
+      </Grid>
+    )
+  })