+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Add `Autocomplete` component.
 - Feature: Add `Field.Autocomplete` component.
+- Fix: Remove `isRequired` of the `ownerName` props on `Comment` component.
+- Fix: Update
+  `Comment`,
+  `CommentForm`,
+  `CommentAvatar` with `styled-components`.
+- Fix: Update `DonationCard` with `styled-components`.
 - Fix: Update
   `TextInput`,
   `TextInputWithButton`,